avformat/hlsenc: Always treat numbers as decimal
8452ece9b77e4348bd6e816ccf378f0f330b2f87 caused a regression: The stream
number is now parsed with strtoll without a fixed basis; as a
consequence, the "010" in a variant stream mapping like "a:010" is now
treated as an octal number (i.e. as eight, not ten). This was not
intended and may break some scripts, so this commit restores the old
behaviour.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>
(cherry picked from commit 19a876fd6973724521dd5e7cc8f8e4683b19eda4)